requirements.txt
langchain==0.1.20
numpy==1.26.4
openai==1.30.1
openpyxl==3.1.2
pandas==2.2.2
# langchain
# numpy
# openai
# openpyxl
# pandas

langchain-openai
langchain-core
langchain-chroma
langchain-community

python-dotenv
unstructured
networkx
python-docx
PyMuPDF

docx2pdf
xlrd
azure-ai-documentintelligence
glob2
psycopg2-binary
